TP#26357: LLBL Change - regen for order item dimension fields decimal conversion

+++ Code/ShipWorks.Data.Model/EntityInterfaces/IOrderItemEntity.cs
@@ -144,21 +144,21 @@ namespace ShipWorks.Data.Model.EntityInterfaces
         /// <summary> The Length property of the Entity OrderItem<br/><br/>
         /// </summary>
         /// <remarks>Mapped on table field: "OrderItem"."Length"<br/>
-        /// Table field type characteristics (type, precision, scale, length): Float, 38, 0, 0<br/>
+        /// Table field type characteristics (type, precision, scale, length): Decimal, 29, 9, 0<br/>
         /// Table field behavior characteristics (is nullable, is PK, is identity): false, false, false</remarks>
-        System.Double Length { get; }
+        System.Decimal Length { get; }
         /// <summary> The Width property of the Entity OrderItem<br/><br/>
         /// </summary>
         /// <remarks>Mapped on table field: "OrderItem"."Width"<br/>
-        /// Table field type characteristics (type, precision, scale, length): Float, 38, 0, 0<br/>
+        /// Table field type characteristics (type, precision, scale, length): Decimal, 29, 9, 0<br/>
         /// Table field behavior characteristics (is nullable, is PK, is identity): false, false, false</remarks>
-        System.Double Width { get; }
+        System.Decimal Width { get; }
         /// <summary> The Height property of the Entity OrderItem<br/><br/>
         /// </summary>
         /// <remarks>Mapped on table field: "OrderItem"."Height"<br/>
-        /// Table field type characteristics (type, precision, scale, length): Float, 38, 0, 0<br/>
+        /// Table field type characteristics (type, precision, scale, length): Decimal, 29, 9, 0<br/>
         /// Table field behavior characteristics (is nullable, is PK, is identity): false, false, false</remarks>
-        System.Double Height { get; }
+        System.Decimal Height { get; }
         
         
         IOrderEntity Order { get; }
