+++ Code/ShipWorks.Data.Model/ReadOnlyEntities/ReadOnlyOrderItemEntity.cs
@@ -193,21 +193,21 @@ namespace ShipWorks.Data.Model.ReadOnlyEntityClasses
         /// <summary> The Length property of the Entity OrderItem<br/><br/>
         /// </summary>
         /// <remarks>Mapped on table field: "OrderItem"."Length"<br/>
-        /// Table field type characteristics (type, precision, scale, length): Float, 38, 0, 0<br/>
+        /// Table field type characteristics (type, precision, scale, length): Decimal, 29, 9, 0<br/>
         /// Table field behavior characteristics (is nullable, is PK, is identity): false, false, false</remarks>
-        public System.Double Length { get; }
+        public System.Decimal Length { get; }
         /// <summary> The Width property of the Entity OrderItem<br/><br/>
         /// </summary>
         /// <remarks>Mapped on table field: "OrderItem"."Width"<br/>
-        /// Table field type characteristics (type, precision, scale, length): Float, 38, 0, 0<br/>
+        /// Table field type characteristics (type, precision, scale, length): Decimal, 29, 9, 0<br/>
         /// Table field behavior characteristics (is nullable, is PK, is identity): false, false, false</remarks>
-        public System.Double Width { get; }
+        public System.Decimal Width { get; }
         /// <summary> The Height property of the Entity OrderItem<br/><br/>
         /// </summary>
         /// <remarks>Mapped on table field: "OrderItem"."Height"<br/>
-        /// Table field type characteristics (type, precision, scale, length): Float, 38, 0, 0<br/>
+        /// Table field type characteristics (type, precision, scale, length): Decimal, 29, 9, 0<br/>
         /// Table field behavior characteristics (is nullable, is PK, is identity): false, false, false</remarks>
-        public System.Double Height { get; }
+        public System.Decimal Height { get; }
         
         
         public IOrderEntity Order { get; }
